split for text that need to be saved

+++ storageSolutions.py
@@ -29,8 +29,8 @@ class storagesolutions:
         if os.path.exists("quizSTG") and os.path.exists("chatSTG"):
             print("folder exist")
         
-        else:
-            pathList = ['quizSTG', 'chatSTG']
+         else:
+                pathList = ['quizSTG', 'chatSTG'] #names of folders on list to iterate through them
             for i in pathList:
                 path = os.path.join(i)
                 os.mkdir(path)
@@ -40,15 +40,55 @@ class storagesolutions:
         file_name = str(subject) + "quiz" + str(grade) #file name, could add datetime module to name the files in a more personalized manner
         file_path = file_name + ".json"
         with open(file_path, 'w') as fp:
-            json.dump(response, fp)
-            os.path.dirname("quizSTG")
-            fp.close()
+            json.dump(response,fp)
+        fp.close
         
     def saveChat(self, chat, subject, grade):
         #
         file_name = 'chat' + str(subject) + str(grade)
         file_path = file_name + ".txt"
         with open(file_path, "w") as fp:
-            os.path.dirname("chatSTG")
-            fp.write(chat)
-            fp.close()
+            fixChat = chat.split() #neeed the necesary parameter to split and parse the chat
+            fp.write(fixChat)
+        fp.close
+    
+
+
+
+'''
+def saveQuiz(response, subject):
+    
+  
+    if os.path.exists("quizSTG"): #checks fro folder to store the quizzes
+
+        filename = "quiz" + str(1) #creates file name, need to find a way to create diff names every time 
+        file_path = filename + ".json"
+        file_dir = os.path.dirname("QuizSTG")
+        with open(file_path, 'w') as fp:
+            json.dump(response, fp)
+    else:    
+        directory = "QuizSTG"
+        # Path
+        path = os.path.join(directory)
+  
+        # Create the directory
+        os.mkdir(path)
+
+def saveChat(response):
+    if os.path.exists("ChatSTG"):
+    
+        count = 1
+        file_path = "Chat" + str(count)
+        count += 1
+        file_dir = os.path.dirname("ChatSTG")
+        with open(file_path, 'w') as fp:
+            json.dump(response, fp)
+    else:    
+        directory = "ChatSTG"
+        # Path
+        path = os.path.join(directory)
+  
+        # Create the directory
+        os.mkdir(path)
+'''
+
